Skip past the initial cut area when starting playback.
Patch from billstuff2001 at sbcglobal dot net

Closes #9538

+++ mythtv/libs/libmythtv/mythplayer.h
@@ -570,7 +570,7 @@ class MPUBLIC MythPlayer
     int jumpchapter;
 
     // Bookmark stuff
-    long long bookmarkseek;
+    uint64_t bookmarkseek;
 
     // Seek
     /// If fftime>0, number of frames to seek forward.
